Clean up
Replacing presenter data models with entities and putting request data models in one module.
Entities would be used as presenters data models at the current implementation because no major different between the data to present and entities for now. However it is good practice to separate the two because they would differ as  user requirements change and system evolves.